simulated annealing prototype, not finished, with test modifications (will be deleted)

+++ simulator/src/main/java/hu/u_szeged/inf/fog/simulator/agent/AgentApplication.java
@@ -20,7 +20,7 @@ public class AgentApplication {
     public static class Resource {
         
         public String name;
-        public Integer cpu;
+        public Double cpu;
         public Long memory;
         public Integer instances;
         public String provider;
